refactor: Add marker modification test and refactor AppViewModel
Added simple marker modification test in info-window component's
modify method. In AppViewModel, moved the createInfoWindowContent
method inside of the createMarker method since that's the only
place it's used. It's functionality is being largely copy-pasted
to within the info-window component's modify method; it's only
being used for testing at the moment, but I should remember to
attempt to reuse the existing code if appropriate.