feat: Make markers filterable by visibility and folder
Also refactored the code to alphabetize the order of properties and
functions. Should've done that in a separate commit, I know. I'll
try to remember for next time.

Added a `visible` property to the Folder model and a generic
`updateVisibility` method to AppViewModel that handles visibility
for markers and folders. The `updateVisibility` method is manually
subscribed to the `sidebar.search` and `sidebar.visibleOnly`
observables, and also called when the map's bounds change via the new
`map.onBoundsChanged` method. The sidebar visibility filtering
observables are no longer manually subscribed to anything else.

Removed the `hidden` custom binding.

In map, added a simple `visibleOnMap` method to support the new
`onBoundsChanged` exposed method.